Remove separate menu folder and file and consolidate in main.js
  * Includes wrapping template definition in new getTemplate() function.
