+++ main.js
@@ -19,24 +19,13 @@ function createWindow () {
   mainWindow.on('closed', function() {
     mainWindow = null;
   });
-// console.log("We're about to require the menu file!");
-//   require('./menu/mainmenu');
-// console.log("We just required the menu file!");
 }
 
 app.on('ready', () => {
-  console.log("1. on ready started");
   createWindow();
-  console.log("2. window created");
-  // var menu = Menu.buildFromTemplate(template);
-  // console.log("3. menu variable created");
-  // Menu.setApplicationMenu(menu);
-  // console.log("4. setApplicationMenu completed");
-console.log("We're about to require the menu file!");
-  require('./menu/mainmenu');
-console.log("We just required the menu file!");
+  const menu = Menu.buildFromTemplate(getTemplate());
+  Menu.setApplicationMenu(menu);
 });
-// app.on('ready', createWindow, console.log("The ready function just ran createWindow!"));
 app.on('window-all-closed', function () {
   if (process.platform !== 'darwin') {
     app.quit();
@@ -47,214 +36,206 @@ app.on('activate', function () {
     createWindow();
   }
 });
-// app.once('ready', function(){
-// console.log("global 1");
-//   // if (Menu.getApplicationMenu())
-//   // console.log("global 2");
-//   //   return;
-// console.log("global 3");
 
-//   var template = [
-//     {
-//       label: 'Edit',
-//       submenu: [
-//         {
-//           label: 'Undo',
-//           accelerator: 'CmdOrCtrl+Z',
-//           role: 'undo'
-//         },
-//         {
-//           label: 'Redo',
-//           accelerator: 'Shift+CmdOrCtrl+Z',
-//           role: 'redo'
-//         },
-//         {
-//           type: 'separator'
-//         },
-//         {
-//           label: 'Cut',
-//           accelerator: 'CmdOrCtrl+X',
-//           role: 'cut'
-//         },
-//         {
-//           label: 'Copy',
-//           accelerator: 'CmdOrCtrl+C',
-//           role: 'copy'
-//         },
-//         {
-//           label: 'Paste',
-//           accelerator: 'CmdOrCtrl+V',
-//           role: 'paste'
-//         },
-//         {
-//           label: 'Select All',
-//           accelerator: 'CmdOrCtrl+A',
-//           role: 'selectall'
-//         },
-//       ]
-//     },
-//     {
-//       label: 'View',
-//       submenu: [
-//         {
-//           label: 'Reload',
-//           accelerator: 'CmdOrCtrl+R',
-//           click: function(item, focusedWindow) {
-//             if (focusedWindow)
-//               focusedWindow.reload();
-//           }
-//         },
-//         {
-//           label: 'Toggle Full Screen',
-//           accelerator: (function() {
-//             if (process.platform == 'darwin')
-//               return 'Ctrl+Command+F';
-//             else
-//               return 'F11';
-//           })(),
-//           click: function(item, focusedWindow) {
-//             if (focusedWindow)
-//               focusedWindow.setFullScreen(!focusedWindow.isFullScreen());
-//           }
-//         },
-//         {
-//           label: 'Toggle Developer Tools',
-//           accelerator: (function() {
-//             if (process.platform == 'darwin')
-//               return 'Alt+Command+I';
-//             else
-//               return 'Ctrl+Shift+I';
-//           })(),
-//           click: function(item, focusedWindow) {
-//             if (focusedWindow)
-//               focusedWindow.toggleDevTools();
-//           }
-//         },
-//       ]
-//     },
-//     {
-//       label: 'Window',
-//       role: 'window',
-//       submenu: [
-//         {
-//           label: 'Minimize',
-//           accelerator: 'CmdOrCtrl+M',
-//           role: 'minimize'
-//         },
-//         {
-//           label: 'Close',
-//           accelerator: 'CmdOrCtrl+W',
-//           role: 'close'
-//         },
-//       ]
-//     },
-//     {
-//       label: 'Help',
-//       role: 'help',
-//       submenu: [
-//         {
-//           label: 'Learn More',
-//           click: function() {
-//             shell.openExternal('https://github.com/nexB/aboutcode-manager');
-//           }
-//         },
-//         {
-//           label: 'Licensing Information',
-//           click: function(item, focusedWindow) {
-//             var win = new BrowserWindow({ frame: true });
-//             win.on('closed', function () { win = null });
-//             win.loadURL('file://' + __dirname + '/attribution.html');
-//             win.show();
+function getTemplate() {
+  const template = [
+    {
+      label: 'Edit',
+      submenu: [
+        {
+          label: 'Undo',
+          accelerator: 'CmdOrCtrl+Z',
+          role: 'undo'
+        },
+        {
+          label: 'Redo',
+          accelerator: 'Shift+CmdOrCtrl+Z',
+          role: 'redo'
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: 'Cut',
+          accelerator: 'CmdOrCtrl+X',
+          role: 'cut'
+        },
+        {
+          label: 'Copy',
+          accelerator: 'CmdOrCtrl+C',
+          role: 'copy'
+        },
+        {
+          label: 'Paste',
+          accelerator: 'CmdOrCtrl+V',
+          role: 'paste'
+        },
+        {
+          label: 'Select All',
+          accelerator: 'CmdOrCtrl+A',
+          role: 'selectall'
+        },
+      ]
+    },
+    {
+      label: 'View',
+      submenu: [
+        {
+          label: 'Reload',
+          accelerator: 'CmdOrCtrl+R',
+          click: function(item, focusedWindow) {
+            if (focusedWindow)
+              focusedWindow.reload();
+          }
+        },
+        {
+          label: 'Toggle Full Screen',
+          accelerator: (function() {
+            if (process.platform == 'darwin')
+              return 'Ctrl+Command+F';
+            else
+              return 'F11';
+          })(),
+          click: function(item, focusedWindow) {
+            if (focusedWindow)
+              focusedWindow.setFullScreen(!focusedWindow.isFullScreen());
+          }
+        },
+        {
+          label: 'Toggle Developer Tools',
+          accelerator: (function() {
+            if (process.platform == 'darwin')
+              return 'Alt+Command+I';
+            else
+              return 'Ctrl+Shift+I';
+          })(),
+          click: function(item, focusedWindow) {
+            if (focusedWindow)
+              focusedWindow.toggleDevTools();
+          }
+        },
+      ]
+    },
+    {
+      label: 'Window',
+      role: 'window',
+      submenu: [
+        {
+          label: 'Minimize',
+          accelerator: 'CmdOrCtrl+M',
+          role: 'minimize'
+        },
+        {
+          label: 'Close',
+          accelerator: 'CmdOrCtrl+W',
+          role: 'close'
+        },
+      ]
+    },
+    {
+      label: 'Help',
+      role: 'help',
+      submenu: [
+        {
+          label: 'Learn More',
+          click: function() {
+            shell.openExternal('https://github.com/nexB/aboutcode-manager');
+          }
+        },
+        {
+          label: 'Licensing Information',
+          click: function(item, focusedWindow) {
+            var win = new BrowserWindow({ frame: true });
+            win.on('closed', function () { win = null });
+            win.loadURL('file://' + __dirname + '/attribution.html');
+            win.show();
 
-//           }
-//         },
-//         {
-//           label: 'Documentation',
-//           click: function() {
-//             shell.openExternal(
-//               `https://github.com/nexB/aboutcode-manager/blob/v${pjson.version}/README.md`
-//             );
-//           }
-//         },
-//         {
-//           label: 'Search Issues',
-//           click: function() {
-//             shell.openExternal('https://github.com/nexB/aboutcode-manager/issues');
-//           }
-//         }
-//       ]
-//     },
-//   ];
-// var pjson = require('./package.json');
+          }
+        },
+        {
+          label: 'Documentation',
+          click: function() {
+            shell.openExternal(
+              `https://github.com/nexB/aboutcode-manager/blob/v${pjson.version}/README.md`
+            );
+          }
+        },
+        {
+          label: 'Search Issues',
+          click: function() {
+            shell.openExternal('https://github.com/nexB/aboutcode-manager/issues');
+          }
+        }
+      ]
+    },
+  ];
+const pjson = require('./package.json');
 
-//   if (process.platform == 'darwin') {
-//     template.unshift({
-//       label: 'AboutCode',
-//       submenu: [
-//         {
-//           label: 'About AboutCode Manager',
-//           click: function(item, focusedWindow) {
-//             var win = new BrowserWindow({ frame: true, width: 250, height: 200 });
-//             win.on('closed', function () { win = null });
-//             win.loadURL('file://' + __dirname + '/about.html');
-//             win.show();
+  if (process.platform == 'darwin') {
+    template.unshift({
+      label: 'AboutCode',
+      submenu: [
+        {
+          label: 'About AboutCode Manager',
+          click: function(item, focusedWindow) {
+            var win = new BrowserWindow({ frame: true, width: 250, height: 200 });
+            win.on('closed', function () { win = null });
+            win.loadURL('file://' + __dirname + '/about.html');
+            win.show();
 
-//           }
-//         },
-//         {
-//           type: 'separator'
-//         },
-//         {
-//           label: `Version ${pjson.version}`,
-//           enabled: false
-//         },
-//         {
-//           type: 'separator'
-//         },
-//         {
-//           label: 'Services',
-//           role: 'services',
-//           submenu: []
-//         },
-//         {
-//           type: 'separator'
-//         },
-//         {
-//           label: 'Hide Electron',
-//           accelerator: 'Command+H',
-//           role: 'hide'
-//         },
-//         {
-//           label: 'Hide Others',
-//           accelerator: 'Command+Alt+H',
-//           role: 'hideothers'
-//         },
-//         {
-//           label: 'Show All',
-//           role: 'unhide'
-//         },
-//         {
-//           type: 'separator'
-//         },
-//         {
-//           label: 'Quit',
-//           accelerator: 'Command+Q',
-//           click: function() { app.quit(); }
-//         },
-//       ]
-//     });
-//     template[3].submenu.push(
-//       {
-//         type: 'separator'
-//       },
-//       {
-//         label: 'Bring All to Front',
-//         role: 'front'
-//       }
-//     );
-//   }
-//   console.log("about to set the menu");
-//   var menu = Menu.buildFromTemplate(template);
-//   Menu.setApplicationMenu(menu);
-//   console.log("we just set the menu");
-// });
+          }
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: `Version ${pjson.version}`,
+          enabled: false
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: 'Services',
+          role: 'services',
+          submenu: []
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: 'Hide Electron',
+          accelerator: 'Command+H',
+          role: 'hide'
+        },
+        {
+          label: 'Hide Others',
+          accelerator: 'Command+Alt+H',
+          role: 'hideothers'
+        },
+        {
+          label: 'Show All',
+          role: 'unhide'
+        },
+        {
+          type: 'separator'
+        },
+        {
+          label: 'Quit',
+          accelerator: 'Command+Q',
+          click: function() { app.quit(); }
+        },
+      ]
+    });
+    template[3].submenu.push(
+      {
+        type: 'separator'
+      },
+      {
+        label: 'Bring All to Front',
+        role: 'front'
+      }
+    );
+  }
+  return template;
+}
